Break out more sub modules in util.
